refactor: ChatRoom 페이지에서 zustand store 활용
